[v4.5] system tests: fix race in kube-play read-only
Use --restart=no on all created containers. Without this, all
containers spin forever and it's impossible to get a
reliable exit status.

As a side effort, clean up tests, make more robust and maintainable.

This is a cherrypick of #18169 onto v4.5, because the flake is
triggering often in CI. The cherrypick applied cleanly.

Signed-off-by: Ed Santiago <santiago@redhat.com>