pkg/rootless: do not use shortcut with --tmpdir
When using --tmpdir for the podman cli we use this as location for the
pause.pid file. However the c shortcut code has no idea about this
option and always assumes XDG_RUNTIME_DIR/libpod/tmp. This can cause us
to join the wrong user namespace which leads to very weird issues as
mounts are missing there.

Fixes #17903

Signed-off-by: Paul Holzinger <pholzing@redhat.com>