Cirrus: Fix aarch64 clone_script 404 errors
Fixes: https://github.com/containers/podman/issues/18040

If the `build_aarch64` task happens to fail for any reason, it will
cause the `curl` command in the `clone_script` for the aarch64 system
test tasks to throw a 404.  This is because the
`local_system_test_aarch64_task` depends on `build` not `build_aarch64`.
As discovered in another issue long ago, the Cirrus API depends on doing
some dependency-resolution magic to function properly.  Fix this by
correcting the dependencies.

Signed-off-by: Chris Evich <cevich@redhat.com>