podman run --authfile do not stat file
No other commands do this, it makes no sense to fail hard if the given
authfile does not exists. We do not even know if we have to pull
anything at that point. I expect the code to error at some later point
if the auth file is really needed but does not exists.

Fixes #18405

Signed-off-by: Paul Holzinger <pholzing@redhat.com>