Skip rhel-release branch unnecessary CI tasks
RHEL release-branches potentially need to be maintained for a very long
time.  Improve reliability and CI-speed by skipping tasks which are not
needed for RHEL.  For example, there will (likely) never be RHEL
releases of the MacOS or Windows versions of podman-remote.  Also,
relocate the `win_installer_task` in the config. to better reflect it's
sequence among CI tasks.

Signed-off-by: Chris Evich <cevich@redhat.com>