Invert state and listening logic
When waiting for a VM to boot up, we check for two indicators to before mounting volumes: is the vm "running" and is the qemu socket listening.  In some cases, presumably race or system pressure, it is possible that the qemu socket may be listening but the vm (and specifically sshd) is not running yet.

Fixes: #17403

[NO NEW TESTS NEEDED]

Signed-off-by: Brent Baude <bbaude@redhat.com>